requirements.txt
edgegrid_python==2.0.1
Flask==3.1.1
flask_httpauth==4.8.0
mohawk==1.1.0
oauthlib==3.2.2
PyJWT==2.10.1
Requests==2.32.4
requests_aws4auth==1.3.1
requests_ntlm==1.3.0
requests_oauthlib==2.0.0
termcolor==3.1.0
